# Represent MARITIME_TRADE as 10-int list