closes #157