@bcollazo 2024-11-11
Hey @zarns Thanks for taking a stab at this! Could you share performance numbers before and after this change? The motivation for this is to allow for faster games / more games per second. I wonder what's the improvement of this.

It might also be interesting to show how does games that involve bots that do a lot of copying (`.copy`) are affected by this change. For example games with the MCTS bot.

Thanks again!