@zarns 2024-11-12
Looks like these changes make the games slightly faster on average with AlphaBetaPlayer/MCTSPlayer/RandomPlayer
- catanatron-play -n 1000
  - before: 43.4 sec
  - after:    42.5 sec
- catanatron-play --players=AB,R,W --num=100
  - before: 309.36 sec
  - after:    285.75 sec
- catanatron-play --players=AB,M --num=10
  - before: 62.27 sec
  - after:    66.8 sec (slightly slower?)
- catanatron-play --players=AB,M,R --num=20
  - before: 224.7 sec
  - after:    206.9 sec (slightly faster on average)